+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.0.0-beta.1
+- [#17](https://github.com/offirgolan/ember-light-table/pull/17) Fixes to table scroll, layout restructuring fixes, and auto set scroll/touch move container if there is a fixed header or footer
+
+## 1.0.0-beta.0
+- [#16](https://github.com/offirgolan/ember-light-table/pull/16) Fixed header and footer support as well as some table layout restructuring.
+
 ## 0.1.9
 - Remove tag-less cell component since it was restricting a bunch of Ember.Component features such as class name bindings, events, etc.
 
@@ -22,7 +28,7 @@
 - Return pushed/inserted rows & columns from public APIs
 
 ## 0.1.3
-- Fixed an issue where cell value was not bounc
+- Fixed an issue where cell value was not bound
 
 ## 0.1.2
 - Ability to push rows using ArrayProxy instances